requirements.txt
beautifulsoup4==4.9.1
PyYAML==5.3.1

